packages/github_client/pyproject.toml
[project]
name = "astra-github-client"
version = "0.1.0"
description = "Astra — thin PyGithub wrapper used by Vault, Bridge-to-Build scaffolder, and Trust Score."
requires-python = ">=3.11,<3.13"
dependencies = [
    "PyGithub>=2.4",
    "pydantic>=2.7,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["astra_github_client"]
